pyproject.toml
[project]
name = "evickastudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=23.7.0",
    "flake8>=6.1.0",
    "flake8-bugbear>=23.9.23",
    "isort>=5.12.0",
    "mypy>=1.4.1",
    "pre-commit>=4.2.0",
    "pytest>=7.3.1",
    "python-dotenv>=0.21.0",
    "requests>=2.25.1",
    "spotipy>=2.22.0",
    "types-requests>=2.32.0.20250515",
]

[tool.black]
line-length = 88
target-version = ['py36']

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
